Calcualte expected descent rate and expected speed more accurately

+++ aman-dman-client/common/src/main/kotlin/AircraftPerformance.kt
@@ -19,7 +19,7 @@ data class AircraftPerformance(
     val cruiseMACH: Float?,
     val cruiseCeiling: Int,
     val cruiseRange: Int?,
-    val initialDescentMACH: Double?,
+    val initialDescentMACH: Float?,
     val initialDescentROD: Int?,
     val descentIAS: Int,
     val descentROD: Int?,
